make the cli more modular
