+++ demo2/README.md
@@ -1,0 +1,9 @@
+# demo2
+
+A custom toolbox for Quantalogic.
+
+## Installation
+
+```bash
+poetry install
+```
