+++ demo2/demo2/tools.py
@@ -1,0 +1,6 @@
+from quantalogic.tools import create_tool
+
+@create_tool
+async def example_tool(param: str) -> str:
+    """An example tool that echoes the input."""
+    return f"Echo: {param}"
