+++ quantalogic/codeact/templates/toolbox/README.md.j2
@@ -1,0 +1,10 @@
+# {{ name }}
+
+A custom toolbox for Quantalogic.
+
+## Installation
+
+```bash
+poetry install
+```
+
